adding AI request to validate crowdfunding

+++ src/main/java/com/donate/donate_app/DTO/CrowdfundingDTO.java
@@ -6,9 +6,9 @@ public class CrowdfundingDTO {
     private String title;
     private String description;
     private Long users_id;
-    private Integer goal_amount;
+    private BigDecimal goal_amount;
 
-    public CrowdfundingDTO(String description, Long users_id, Integer goal_amount, String title) {
+    public CrowdfundingDTO(String description, Long users_id, BigDecimal goal_amount, String title) {
         this.description = description;
         this.users_id = users_id;
         this.goal_amount = goal_amount;
@@ -34,11 +34,11 @@ public class CrowdfundingDTO {
         this.users_id = users_id;
     }
 
-    public Integer getGoal_amount() {
+    public BigDecimal getGoal_amount() {
         return goal_amount;
     }
 
-    public void setGoal_amount(Integer goal_amount) {
+    public void setGoal_amount(BigDecimal goal_amount) {
         this.goal_amount = goal_amount;
     }
